requirements.txt
numpy==1.26.4
pandas==2.2.2
pyarrow==15.0.2
streamlit
openai
langchain
langchain-community
langchain-openai
openai
python-dotenv
pymupdf
python-docx
faiss-cpu
docx2txt
pypdf
chromadb
tiktoken
unstructured
unstructured[pdf]
